[FIX] javax.validation -> jakarta.validation 의존성 변경 및 UserDupCheckvalidator 적용
