+++ src/main/java/com/kujproject/kuj/dto/SignUpReqDto.java
@@ -3,15 +3,17 @@ package com.kujproject.kuj.dto;
 import com.kujproject.kuj.validation.user.EmailDupCheckValidator;
 import com.kujproject.kuj.validation.user.IdDupCheckValidator;
 import jakarta.persistence.Column;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotEmpty;
 import lombok.Data;
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotEmpty;
+
 
 @Data
 public class SignUpReqDto {
     @Column(unique = true)
-    @IdDupCheckValidator
+
     @NotEmpty(message = "ID는 필수 항목입니다.")
+    @IdDupCheckValidator
     String userId;
 
     @NotEmpty(message = "비밀번호는 필수 항목입니다.")
@@ -24,8 +26,8 @@ public class SignUpReqDto {
     String userName;
 
     @Column(unique = true)
-    @EmailDupCheckValidator
     @Email(message = "유효한 이메일을 입력해주세요.")
+    @EmailDupCheckValidator
     String email;
 
     @Column(unique = true)
