+++ src/main/java/com/kujproject/kuj/dto/UpdatePasswordReqDto.java
@@ -1,8 +1,8 @@
 package com.kujproject.kuj.dto;
 
+import jakarta.validation.constraints.NotEmpty;
 import lombok.Data;
 
-import javax.validation.constraints.NotEmpty;
 
 @Data
 public class UpdatePasswordReqDto {
