+++ src/main/java/com/kujproject/kuj/validation/user/EmailDupCheckValidator.java
@@ -1,7 +1,8 @@
 package com.kujproject.kuj.validation.user;
 
-import javax.validation.Constraint;
-import javax.validation.Payload;
+import jakarta.validation.Constraint;
+import jakarta.validation.Payload;
+
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -9,7 +10,7 @@ import java.lang.annotation.Target;
 
 @Target({ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
-//@Constraint(validatedBy = )
+@Constraint(validatedBy = UserDupCheckValidator.EmailDupCheck.class)
 public @interface EmailDupCheckValidator {
     String message() default "이미 등록 된 이메일입니다.";
