+++ src/main/java/com/kujproject/kuj/validation/user/IdDupCheckValidator.java
@@ -1,6 +1,8 @@
 package com.kujproject.kuj.validation.user;
 
-import javax.validation.Payload;
+import jakarta.validation.Constraint;
+import jakarta.validation.Payload;
+
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -8,7 +10,7 @@ import java.lang.annotation.Target;
 
 @Target({ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
-//@Constraint(validatedBy = )
+@Constraint(validatedBy = UserDupCheckValidator.IdDupCheck.class)
 public @interface IdDupCheckValidator {
     String message() default "이미 사용 중인 ID입니다.";
