+++ src/main/java/com/kujproject/kuj/web/controller/UserController.java
@@ -2,18 +2,25 @@ package com.kujproject.kuj.web.controller;
 
 import com.kujproject.kuj.domain.service.UserService;
 import com.kujproject.kuj.dto.*;
+import com.kujproject.kuj.validation.user.IdDupCheckValidator;
+import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.validation.BindingResult;
+import org.springframework.validation.FieldError;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.Valid;
+
+import java.util.HashMap;
+import java.util.Map;
 import java.util.Optional;
+
 
 @RestController
 public class UserController {
-    private UserService userService;
+    private final UserService userService;
     private final PasswordEncoder passwordEncoder;
 
     //자동 주입(spring 4.3 < ver)
@@ -23,11 +30,19 @@ public class UserController {
     }
     
     @PostMapping("/user")
-    public ResponseEntity<SignUpReqDto> signUp(@Valid @RequestBody SignUpReqDto signUpReqDto, BindingResult bindingResult) {
+    public ResponseEntity<?> signUp(@Valid @RequestBody SignUpReqDto signUpReqDto, BindingResult bindingResult) {
         // bindingResult 에러 검출
-        
+        if(bindingResult.hasErrors()) {
+            Map<String, String> errorMap = new HashMap<>();
+
+            for(FieldError error : bindingResult.getFieldErrors()) {
+                errorMap.put(error.getField(), error.getDefaultMessage());
+            }
+
+            return ResponseEntity.badRequest().body(errorMap);
+        }
+        System.out.println(23321);
         userService.signUp(signUpReqDto);
-        
         return ResponseEntity.status(HttpStatus.CREATED).body(signUpReqDto);
     }
 
@@ -46,18 +61,36 @@ public class UserController {
     }
 
     @PatchMapping("/user/{id}/email")
-    public ResponseEntity<UpdateEmailReqDto> updateEmail(@PathVariable String id,
+    public ResponseEntity<?> updateEmail(@PathVariable String id,
                                                          @Valid @RequestBody UpdateEmailReqDto updateEmailReqDto, BindingResult bindingResult) {
         // bindingResult 에러 검출
+        if(bindingResult.hasErrors()) {
+            Map<String, String> errorMap = new HashMap<>();
+
+            for(FieldError error : bindingResult.getFieldErrors()) {
+                errorMap.put(error.getField(), error.getDefaultMessage());
+            }
+
+            return ResponseEntity.badRequest().body(errorMap);
+        }
 
         userService.updateEmail(id, updateEmailReqDto);
         return ResponseEntity.ok().body(updateEmailReqDto);
     }
 
     @PatchMapping("/user/{id}/passwd")
-    public ResponseEntity<UpdatePasswordReqDto> updateEmail(@PathVariable String id,
+    public ResponseEntity<?> updateEmail(@PathVariable String id,
                                                             @Valid @RequestBody UpdatePasswordReqDto updatePasswordReqDto, BindingResult bindingResult) {
         // bindingResult 에러 검출
+        if(bindingResult.hasErrors()) {
+            Map<String, String> errorMap = new HashMap<>();
+
+            for(FieldError error : bindingResult.getFieldErrors()) {
+                errorMap.put(error.getField(), error.getDefaultMessage());
+            }
+
+            return ResponseEntity.badRequest().body(errorMap);
+        }
 
         boolean isUpdated = userService.updatePassword(id,updatePasswordReqDto);
 
@@ -70,9 +103,18 @@ public class UserController {
     }
 
     @PatchMapping("/user/{id}/profile")
-    public ResponseEntity<UpdateProfileReqDto> updateProfile(@PathVariable String id,
+    public ResponseEntity<?> updateProfile(@PathVariable String id,
                                                              @Valid @RequestBody UpdateProfileReqDto updateProfileReqDto, BindingResult bindingResult) {
         // bindingResult 에러 검출
+        if(bindingResult.hasErrors()) {
+            Map<String, String> errorMap = new HashMap<>();
+
+            for(FieldError error : bindingResult.getFieldErrors()) {
+                errorMap.put(error.getField(), error.getDefaultMessage());
+            }
+
+            return ResponseEntity.badRequest().body(errorMap);
+        }
 
         userService.updateUserProfile(id, updateProfileReqDto);
         return ResponseEntity.ok().body(updateProfileReqDto);
